+++ cursed/deobfuscator/.gitignore
@@ -1,5 +1,15 @@
-*
-!ast.js
-!ast_utils.js
-!package.json
-!.gitignore
+# Test folders
+old/
+tests/
+output/*
+!output/.gitkeep
+
+# Node JS
+node_modules
+package-lock.json
+
+# test run script
+run.sh
+
+# temp todo
+TODO
