+++ cursed/deobfuscator/ast/r6d1qt.js
@@ -1,0 +1,89 @@
+const { UnaryExpressionToValue, valueToNode } = require('../ast_utils');
+const { bbystealsurheartwithlovebrunxkd$2kwqqf, bbystealsurheartwithlovebrunxkd$ffoxnc, bbystealsurheartwithlovebrunxkd$ucypl } = require('../vars/vars');
+const { bbystealsurheartwithlovebrunxkd$ukyoii } = require('./ukyoii')
+
+function bbystealsurheartwithlovebrunxkd$r6d1qt(bbystealsurheartwithlovebrunxkd$sbzfdo, bbystealsurheartwithlovebrunxkd$oth1jj, bbystealsurheartwithlovebrunxkd$js15qs, bbystealsurheartwithlovebrunxkd$te6clf = bbystealsurheartwithlovebrunxkd$z1hel, bbystealsurheartwithlovebrunxkd$ige25 = bbystealsurheartwithlovebrunxkd$ucypl) {
+    if (bbystealsurheartwithlovebrunxkd$js15qs) {
+        return bbystealsurheartwithlovebrunxkd$oth1jj[bbystealsurheartwithlovebrunxkd$ucypl[bbystealsurheartwithlovebrunxkd$js15qs]] = bbystealsurheartwithlovebrunxkd$r6d1qt(bbystealsurheartwithlovebrunxkd$sbzfdo, bbystealsurheartwithlovebrunxkd$oth1jj)
+    } else if (bbystealsurheartwithlovebrunxkd$oth1jj) {
+        [bbystealsurheartwithlovebrunxkd$ige25, bbystealsurheartwithlovebrunxkd$oth1jj] = [bbystealsurheartwithlovebrunxkd$te6clf(bbystealsurheartwithlovebrunxkd$ige25), bbystealsurheartwithlovebrunxkd$sbzfdo || bbystealsurheartwithlovebrunxkd$js15qs]
+    }
+    return bbystealsurheartwithlovebrunxkd$oth1jj ? bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$ige25[bbystealsurheartwithlovebrunxkd$oth1jj]] : bbystealsurheartwithlovebrunxkd$ucypl[bbystealsurheartwithlovebrunxkd$sbzfdo] || (bbystealsurheartwithlovebrunxkd$js15qs = (bbystealsurheartwithlovebrunxkd$ige25[bbystealsurheartwithlovebrunxkd$sbzfdo], bbystealsurheartwithlovebrunxkd$te6clf), bbystealsurheartwithlovebrunxkd$ucypl[bbystealsurheartwithlovebrunxkd$sbzfdo] = bbystealsurheartwithlovebrunxkd$js15qs(bbystealsurheartwithlovebrunxkd$ffoxnc[bbystealsurheartwithlovebrunxkd$sbzfdo]))
+}
+
+// linked to r6d1qt
+function bbystealsurheartwithlovebrunxkd$z1hel(...bbystealsurheartwithlovebrunxkd$sbzfdo) {
+    var bbystealsurheartwithlovebrunxkd$oth1jj = function() {
+        return function(bbystealsurheartwithlovebrunxkd$sbzfdo) {
+            return bbystealsurheartwithlovebrunxkd$2kwqqf[bbystealsurheartwithlovebrunxkd$sbzfdo + -175]
+        }
+    }();
+    ~(bbystealsurheartwithlovebrunxkd$sbzfdo.length = bbystealsurheartwithlovebrunxkd$ukyoii(708), bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$ukyoii(694)] = -bbystealsurheartwithlovebrunxkd$ukyoii(910), bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$oth1jj(937)] - (bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$ukyoii(694)] - bbystealsurheartwithlovebrunxkd$ukyoii(708))] = '', bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$ukyoii(694)] - (bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$ukyoii(694)] - (bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$oth1jj(937)] - bbystealsurheartwithlovebrunxkd$ukyoii(694))] - bbystealsurheartwithlovebrunxkd$oth1jj(869))] = bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$ukyoii(694)] - -bbystealsurheartwithlovebrunxkd$oth1jj(882)], bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$ukyoii(694)] - -bbystealsurheartwithlovebrunxkd$oth1jj(867)] - (bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$oth1jj(937)] - -bbystealsurheartwithlovebrunxkd$ukyoii(624)] - bbystealsurheartwithlovebrunxkd$oth1jj(937))] - -(bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$ukyoii(694)] - -bbystealsurheartwithlovebrunxkd$ukyoii(83))] - (bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$oth1jj(937)] - -bbystealsurheartwithlovebrunxkd$ukyoii(624)] - -bbystealsurheartwithlovebrunxkd$oth1jj(867)] - -bbystealsurheartwithlovebrunxkd$ukyoii(624)] - (bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$ukyoii(694)] - (bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$oth1jj(937)] - bbystealsurheartwithlovebrunxkd$ukyoii(-67))))] = bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$oth1jj(937)] - -(bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$ukyoii(694)] - -bbystealsurheartwithlovebrunxkd$ukyoii(624)] - (bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$oth1jj(937)] - -bbystealsurheartwithlovebrunxkd$oth1jj(867)] - bbystealsurheartwithlovebrunxkd$ukyoii(910)))].substring(bbystealsurheartwithlovebrunxkd$oth1jj(951), bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$ukyoii(694)] - -bbystealsurheartwithlovebrunxkd$oth1jj(867)] - (bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$ukyoii(694)] - bbystealsurheartwithlovebrunxkd$ukyoii(694))] - (bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$ukyoii(694)] - (bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$oth1jj(937)] - (bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$ukyoii(694)] - -bbystealsurheartwithlovebrunxkd$oth1jj(867)] - -bbystealsurheartwithlovebrunxkd$oth1jj(867)] - bbystealsurheartwithlovebrunxkd$ukyoii(694))] - (bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$ukyoii(694)] - bbystealsurheartwithlovebrunxkd$ukyoii(-67))))].length - bbystealsurheartwithlovebrunxkd$oth1jj(951)), bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$oth1jj(937)] - (bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$oth1jj(937)] - bbystealsurheartwithlovebrunxkd$ukyoii(633))] = bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$oth1jj(937)] - -bbystealsurheartwithlovebrunxkd$oth1jj(867)] - (bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$ukyoii(694)] - -bbystealsurheartwithlovebrunxkd$oth1jj(958)), bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$oth1jj(937)] - -bbystealsurheartwithlovebrunxkd$ukyoii(624)] - -bbystealsurheartwithlovebrunxkd$oth1jj(1061)] = bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$ukyoii(694)] - -bbystealsurheartwithlovebrunxkd$oth1jj(208)] - (bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$ukyoii(633)] - (bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$ukyoii(633)] - (bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$ukyoii(694)] - -bbystealsurheartwithlovebrunxkd$ukyoii(624)] - -bbystealsurheartwithlovebrunxkd$ukyoii(624)] - -bbystealsurheartwithlovebrunxkd$oth1jj(867)] - (bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$ukyoii(694)] - -bbystealsurheartwithlovebrunxkd$oth1jj(1523)))] - -bbystealsurheartwithlovebrunxkd$oth1jj(958)))].split('\u002c'));
+    for (var bbystealsurheartwithlovebrunxkd$js15qs = bbystealsurheartwithlovebrunxkd$oth1jj(176); bbystealsurheartwithlovebrunxkd$js15qs < bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$ukyoii(694)] - (bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$oth1jj(937)] - -bbystealsurheartwithlovebrunxkd$ukyoii(624)] - -bbystealsurheartwithlovebrunxkd$oth1jj(208)] - bbystealsurheartwithlovebrunxkd$ukyoii(641))].length; bbystealsurheartwithlovebrunxkd$js15qs += bbystealsurheartwithlovebrunxkd$ukyoii(710)) {
+        var bbystealsurheartwithlovebrunxkd$te6clf = function() {
+            return function(bbystealsurheartwithlovebrunxkd$sbzfdo) {
+                return bbystealsurheartwithlovebrunxkd$ukyoii(bbystealsurheartwithlovebrunxkd$sbzfdo + 227)
+            }
+        }();
+        bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$oth1jj(876)] - -bbystealsurheartwithlovebrunxkd$te6clf(510)] = [bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$te6clf(467)] - -bbystealsurheartwithlovebrunxkd$te6clf(591)][bbystealsurheartwithlovebrunxkd$js15qs], bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$oth1jj(937)] - -bbystealsurheartwithlovebrunxkd$ukyoii(-35)] - -bbystealsurheartwithlovebrunxkd$oth1jj(934)] - -bbystealsurheartwithlovebrunxkd$oth1jj(208)] - -bbystealsurheartwithlovebrunxkd$ukyoii(707)][bbystealsurheartwithlovebrunxkd$js15qs + bbystealsurheartwithlovebrunxkd$te6clf(481)]];
+        var [bbystealsurheartwithlovebrunxkd$ige25, bbystealsurheartwithlovebrunxkd$orvr1h] = bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$oth1jj(937)] - -bbystealsurheartwithlovebrunxkd$te6clf(-262)] - -bbystealsurheartwithlovebrunxkd$oth1jj(1523)] - -bbystealsurheartwithlovebrunxkd$te6clf(510)].map(Number);
+        while (bbystealsurheartwithlovebrunxkd$orvr1h) {
+            typeof(bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$oth1jj(876)] - -bbystealsurheartwithlovebrunxkd$ukyoii(691)] - -bbystealsurheartwithlovebrunxkd$oth1jj(790)] += String.fromCharCode(bbystealsurheartwithlovebrunxkd$ige25 >> bbystealsurheartwithlovebrunxkd$oth1jj(958) * (bbystealsurheartwithlovebrunxkd$orvr1h & bbystealsurheartwithlovebrunxkd$oth1jj(946)) & bbystealsurheartwithlovebrunxkd$oth1jj(858)), bbystealsurheartwithlovebrunxkd$orvr1h >>= bbystealsurheartwithlovebrunxkd$ukyoii(735))
+        }
+    }
+    if (bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$oth1jj(937)] - -bbystealsurheartwithlovebrunxkd$oth1jj(208)] > bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$oth1jj(876)] - -bbystealsurheartwithlovebrunxkd$oth1jj(908)) {
+        var bbystealsurheartwithlovebrunxkd$sl809 = function() {
+            return function(bbystealsurheartwithlovebrunxkd$sbzfdo) {
+                return bbystealsurheartwithlovebrunxkd$ukyoii(bbystealsurheartwithlovebrunxkd$sbzfdo + 122)
+            }
+        }();
+        return bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sl809(511)] - (bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$ukyoii(633)] - bbystealsurheartwithlovebrunxkd$sl809(551))]
+    } else {
+        return bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$sbzfdo[bbystealsurheartwithlovebrunxkd$oth1jj(876)] - -bbystealsurheartwithlovebrunxkd$ukyoii(623)].replace(/~/g, '')
+    }
+}
+
+function ast_r6d1qt(node) {
+    // if (node.arguments.length > 2) { // testing check
+    //     console.log('alert');
+    //     return node;
+    // }
+    if (node.parent && node.parent.type === 'VariableDeclarator') {
+        const arg = node.arguments[1];
+        let value = null;
+        switch (arg.type) {
+            case 'ArrayExpression':
+                value = arg.elements[0];
+                switch (value.type) {
+                    case 'Literal':
+                        value = value.value;
+                        break;
+                    case 'UnaryExpression':
+                        value = UnaryExpressionToValue(value);
+                        break;
+                    default:
+                        // call expression that are not yet evaluated
+                        // console.error(value.type);
+                        return node;
+                }
+                break;
+            case 'Literal':
+                value = arg.value;
+                break;
+            case 'UnaryExpression':
+                value = UnaryExpressionToValue(arg);
+                break;
+            default:
+                // TODO : complete other expressions (many evaluated before)
+                // console.error(arg.type);
+                return node;
+        }
+        let resNode = valueToNode(node, bbystealsurheartwithlovebrunxkd$r6d1qt.apply(this, [value]));
+        if (!resNode)
+            return node;
+        return resNode;
+    }
+    return node
+}
+
+module.exports = { ast_r6d1qt, bbystealsurheartwithlovebrunxkd$r6d1qt }
