+++ cursed/deobfuscator/ast/ukyoii.js
@@ -1,0 +1,18 @@
+const { valueToNode, getValueArg } = require("../ast_utils");
+const { bbystealsurheartwithlovebrunxkd$2kwqqf } = require("../vars/vars");
+
+// This "ukyoii" function is repeated 11000 times to retrieve via index elements from the big arrays below
+function bbystealsurheartwithlovebrunxkd$ukyoii(val) {
+    return bbystealsurheartwithlovebrunxkd$2kwqqf[val + 68];
+};
+
+function ast_ukyoii(node) {
+    const rawArg = node.arguments[0];
+    const nbRes = getValueArg(rawArg);
+    if (nbRes == null)
+        return node;
+    const res = bbystealsurheartwithlovebrunxkd$ukyoii(nbRes);
+    return valueToNode(node, res);
+}
+
+module.exports = { ast_ukyoii, bbystealsurheartwithlovebrunxkd$ukyoii }
